Refactor `app/components/chat`
