--- src/App/components/Chat/Chat.module.css
@@ -1,159 +1,0 @@
-.chat_wrapper,
-.chat_wrapper_active {
-    width: 360px;
-    padding: 1rem;
-
-    height: 400px;
-    position: absolute;
-    background: var(--dark3);
-    border-radius: 4px;
-    box-shadow: 0px 35px 20px rgba(0, 0, 0, 0.3);
-    transition: all 0.6s cubic-bezier(0.68, -0.55, 0.265, 1.55);
-    right: 10px;
-
-    /* flex-direction: column;
-
-    justify-content: space-between; */
-
-    display: grid;
-    grid-template-rows: 24px auto 24px;
-}
-
-.welcome_back {
-    width: 255px;
-    padding: 1rem;
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    position: absolute;
-    right: 50px;
-    bottom: 60px;
-    border-radius: var(--border-radius);
-    background: var(--dark3);
-    color: var(--text-grey-white);
-    cursor: pointer;
-    transition: all 500ms ease-in-out;
-}
-
-.chat_wrapper_full {
-    height: calc(100vh - 5rem);
-    width: 100%;
-    padding: 1rem;
-
-    background: var(--dark3);
-    border-radius: 4px;
-    box-shadow: 0px 35px 20px rgba(0, 0, 0, 0.3);
-    transition: all 0.6s cubic-bezier(0.68, -0.55, 0.265, 1.55);
-    right: 10px;
-
-    grid-template-rows: 24px auto 24px;
-    opacity: 1;
-}
-
-.chat_wrapper {
-    bottom: 0px;
-
-    opacity: 0;
-    pointer-events: none;
-}
-
-.chat_wrapper_active {
-    opacity: 1;
-    bottom: 85px;
-    pointer-events: auto;
-}
-.icon {
-    transition: all 0.4s ease;
-}
-
-.chat_button,
-.chat_wrapper,
-.chat_wrapper_active {
-    display: none;
-}
-
-.chat_header {
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-
-    font-weight: 200;
-    font-size: 18px;
-    line-height: 22px;
-
-    color: var(--text-grey-white);
-}
-
-.chat_header_right {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 8px;
-}
-
-.chat_header svg {
-    cursor: pointer;
-}
-
-.avatar {
-    width: 24px;
-    height: 24px;
-    background: var(--accent-secondary);
-    border-radius: 50%;
-}
-/* chat input */
-.input_container {
-    background: var(--dark2);
-    padding: 0 8px;
-    display: grid;
-    grid-template-columns: auto 60px;
-    align-items: center;
-    height: 22px;
-}
-.input_container input::placeholder {
-    color: var(--text-grey-dark);
-}
-
-.input_container input {
-    background: transparent;
-    border-radius: 4px;
-    color: var(--text-grey-white);
-    font-weight: 300;
-    font-size: 12px;
-    line-height: 15px;
-
-    outline: none;
-    border: none;
-}
-
-.input_container .input_right {
-    display: flex;
-    justify-content: flex-end;
-}
-
-.name {
-    color: var(--accent-secondary);
-
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-}
-
-/* messages display */
-
-.messages_container {
-    width: 100%;
-
-    /* background: var(--dark2); */
-
-    overflow-y: scroll;
-}
-.messages_container::-webkit-scrollbar {
-    display: none;
-}
-@media only screen and (min-width: 720px) {
-    .chat_button,
-    .chat_wrapper,
-    .chat_wrapper_active,
-    .chat_wrapper_full {
-        display: grid;
-    }
-}
